break out classes and static methods
