+++ lib/swid/frontend/dart/dartClassOrMixinOrClassTypAliasDeclaration.freezed.dart
@@ -1,7 +1,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
 
-part of 'dartClass.dart';
+part of 'dartClassOrMixinOrClassTypAliasDeclaration.dart';
 
 // **************************************************************************
 // FreezedGenerator
