+++ lib/swid/frontend/dart/dartFrontend.dart
@@ -6,11 +6,12 @@ import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/file_system/file_system.dart' hide File;
 import 'package:analyzer/source/line_info.dart';
 import 'package:analyzer/src/dart/ast/ast.dart';
+import 'package:hydro_sdk/swid/frontend/dart/dartClassOrMixinOrClassTypAliasDeclaration.dart';
+import 'package:hydro_sdk/swid/frontend/dart/swidClassFromDartClassOrMixinOrClassTypAliasDeclaration.dart';
 import 'package:meta/meta.dart';
 import 'package:surveyor/src/driver.dart';
 import 'package:surveyor/src/visitors.dart';
 
-import 'package:hydro_sdk/swid/frontend/dart/dartClass.dart';
 import 'package:hydro_sdk/swid/frontend/swidFrontend.dart';
 import 'package:hydro_sdk/swid/ir/swidClass.dart';
 import 'package:hydro_sdk/swid/ir/swidDeclarationModifiers.dart';
@@ -95,11 +96,10 @@ class _SwidVisitor extends RecursiveAstVisitor
   @override
   void visitClassDeclaration(ClassDeclaration node) {
     if (node.nativeClause == null) {
-      var res =
-          DartClass.swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
-              dartClassOrMixinOrClassTypAliasDeclaration:
-                  DartClassOrMixinOrClassTypAliasDeclaration
-                      .fromClassDeclaration(classDeclaration: node));
+      var res = swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
+          dartClassOrMixinOrClassTypAliasDeclaration:
+              DartClassOrMixinOrClassTypAliasDeclaration.fromClassDeclaration(
+                  classDeclaration: node));
 
       if (res != null) {
         res = SwidClass.clone(
@@ -248,7 +248,7 @@ class _SwidVisitor extends RecursiveAstVisitor
 
   @override
   void visitClassTypeAlias(ClassTypeAlias node) {
-    var res = DartClass.swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
+    var res = swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
         dartClassOrMixinOrClassTypAliasDeclaration:
             DartClassOrMixinOrClassTypAliasDeclaration.fromClassTypeAlias(
                 classTypeAlias: node));
@@ -275,11 +275,10 @@ class _SwidVisitor extends RecursiveAstVisitor
     if (node.name.name == "Diagnosticable") {
       print(node.name.name);
 
-      var res =
-          DartClass.swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
-              dartClassOrMixinOrClassTypAliasDeclaration:
-                  DartClassOrMixinOrClassTypAliasDeclaration
-                      .fromMixinDeclaration(mixinDeclaration: node));
+      var res = swidClassFromDartClassOrMixinOrClassTypAliasDeclaration(
+          dartClassOrMixinOrClassTypAliasDeclaration:
+              DartClassOrMixinOrClassTypAliasDeclaration.fromMixinDeclaration(
+                  mixinDeclaration: node));
 
       if (res != null) {
         classes.add(res);
