+++ lib/swid/frontend/dart/swidClassFromDartClassOrMixinOrClassTypAliasDeclaration.dart
@@ -1,21 +1,13 @@
 import 'package:analyzer/dart/ast/ast.dart'
     show
-        ClassDeclaration,
         FieldDeclaration,
         VariableDeclaration,
         VariableDeclarationList,
-        TypeParameterList,
-        ClassTypeAlias,
-        MixinDeclaration,
-        SimpleIdentifier,
-        ImplementsClause,
-        WithClause,
-        TypeName;
+        TypeParameterList;
 
-import 'package:analyzer/dart/ast/syntactic_entity.dart';
-import 'package:analyzer/dart/element/element.dart' show ClassElement;
 import 'package:analyzer/dart/element/type.dart' show InterfaceType;
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:hydro_sdk/swid/frontend/dart/dartClassOrMixinOrClassTypAliasDeclaration.dart';
 import 'package:meta/meta.dart';
 
 import 'package:hydro_sdk/swid/ir/swidClass.dart';
@@ -31,71 +23,7 @@ import 'package:hydro_sdk/swid/ir/util/narrowModifierProducer.dart';
 import 'package:analyzer/src/dart/ast/ast.dart'
     show ConstructorDeclarationImpl, MethodDeclarationImpl;
 
-part 'dartClass.freezed.dart';
-
-@freezed
-abstract class DartClassOrMixinOrClassTypAliasDeclaration
-    with _$DartClassOrMixinOrClassTypAliasDeclaration {
-  const factory DartClassOrMixinOrClassTypAliasDeclaration.fromClassDeclaration({
-    @required ClassDeclaration classDeclaration,
-  }) = _$DartClassOrMixinOrClassTypAliasDeclarationFromClassDeclaration;
-
-  const factory DartClassOrMixinOrClassTypAliasDeclaration.fromMixinDeclaration({
-    @required MixinDeclaration mixinDeclaration,
-  }) = _$DartClassOrMixinOrClassTypAliasDeclarationFromMixinDeclaration;
-
-  const factory DartClassOrMixinOrClassTypAliasDeclaration.fromClassTypeAlias({
-    @required ClassTypeAlias classTypeAlias,
-  }) = _$DartClassOrMixinOrClassTypAliasDeclarationFromClassTypeAlias;
-}
-
-extension DartClassOrMixinOrClassTypAliasDeclarationMethods
-    on DartClassOrMixinOrClassTypAliasDeclaration {
-  Iterable<SyntacticEntity> get childEntities => this.when(
-        fromClassDeclaration: (val) => val.childEntities,
-        fromMixinDeclaration: (val) => val.childEntities,
-        fromClassTypeAlias: (val) => val.childEntities,
-      );
-
-  SimpleIdentifier get name => this.when(
-        fromClassDeclaration: (val) => val.name,
-        fromMixinDeclaration: (val) => val.name,
-        fromClassTypeAlias: (val) => val.name,
-      );
-
-  TypeName get superClass => this.when(
-        fromClassDeclaration: (val) => val.extendsClause?.superclass,
-        fromMixinDeclaration: (_) => null,
-        fromClassTypeAlias: (val) => val.superclass,
-      );
-
-  ImplementsClause get implementsClause => this.when(
-        fromClassDeclaration: (val) => val.implementsClause,
-        fromMixinDeclaration: (val) => val.implementsClause,
-        fromClassTypeAlias: (val) => val.implementsClause,
-      );
-
-  WithClause get withClause => this.when(
-        fromClassDeclaration: (val) => val.withClause,
-        fromMixinDeclaration: (_) => null,
-        fromClassTypeAlias: (val) => val.withClause,
-      );
-
-  ClassElement get declaredElement => this.when(
-        fromClassDeclaration: (val) => val.declaredElement,
-        fromMixinDeclaration: (val) => val.declaredElement,
-        fromClassTypeAlias: (val) => val.declaredElement,
-      );
-
-  bool get isMixin => this.when(
-        fromClassDeclaration: (_) => false,
-        fromMixinDeclaration: (_) => true,
-        fromClassTypeAlias: (_) => false,
-      );
-}
-
-class DartClass {
-  static SwidClass swidClassFromDartClassOrMixinOrClassTypAliasDeclaration({
+SwidClass swidClassFromDartClassOrMixinOrClassTypAliasDeclaration({
     @required
         DartClassOrMixinOrClassTypAliasDeclaration
             dartClassOrMixinOrClassTypAliasDeclaration,
@@ -254,4 +182,3 @@ class DartClass {
           orElse: () => null,
         )));
   }
-}
